moved ticker tests to own module
